+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 #### Improvements
 * Fix #6863: ensuring SerialExecutor does not throw RejectedExecutionException to prevent unnecessary error logs
 * Fix #6763: (crd-generator) YAML output customization
+* Fix #6880: LogWatch interface provides listeners on close stream event 
 
 #### Dependency Upgrade